rails/pyproject.toml
[project]
name = "rails"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.2"
dependencies = [
    "paho-mqtt>=2.1.0",
    "python-dotenv>=1.2.1",
    "websockets>=16.0",
]
